requirements.txt
djangorestframework
python-dotenv
pytest
pytest-django